+++ src/UglyToad.PdfPig.Tests/Writer/PdfDocumentBuilderTests.cs
@@ -250,10 +250,10 @@
             page1.AddText("incididunt ut labore et dolore magna aliqua.", 9, new PdfPoint(30, topLine.Y - letters.Max(x => x.GlyphRectangle.Height) - 5), font);
 
             var page2Letters = page2.AddText("The very hungry caterpillar ate all the apples in the garden.", 12, topLine, font);
-            var left = page2Letters[0].GlyphRectangle.Left;
-            var bottom = page2Letters.Min(x => x.GlyphRectangle.Bottom);
-            var right = page2Letters[page2Letters.Count - 1].GlyphRectangle.Right;
-            var top = page2Letters.Max(x => x.GlyphRectangle.Top);
+            var left = (decimal)page2Letters[0].GlyphRectangle.Left;
+            var bottom = (decimal)page2Letters.Min(x => x.GlyphRectangle.Bottom);
+            var right = (decimal)page2Letters[page2Letters.Count - 1].GlyphRectangle.Right;
+            var top = (decimal)page2Letters.Max(x => x.GlyphRectangle.Top);
             page2.SetStrokeColor(10, 250, 69);
             page2.DrawRectangle(new PdfPoint(left, bottom), right - left, top - bottom);
