+++ src/UglyToad.PdfPig/Fonts/TrueType/TrueTypeFontProgram.cs
@@ -60,7 +60,6 @@
                     }
                 }
             }
-
         }
 
         public bool TryGetBoundingBox(int characterIdentifier, out PdfRectangle boundingBox) => TryGetBoundingBox(characterIdentifier, null, out boundingBox);
@@ -93,10 +92,10 @@
             return true;
         }
 
-        public bool TryGetBoundingAdvancedWidth(int characterIdentifier, out decimal width) => TryGetBoundingAdvancedWidth(characterIdentifier, null, out width);
-        public bool TryGetBoundingAdvancedWidth(int characterIdentifier, Func<int, int?> characterCodeToGlyphId, out decimal width)
+        public bool TryGetBoundingAdvancedWidth(int characterIdentifier, out double width) => TryGetBoundingAdvancedWidth(characterIdentifier, null, out width);
+        public bool TryGetBoundingAdvancedWidth(int characterIdentifier, Func<int, int?> characterCodeToGlyphId, out double width)
         {
-            width = 0m;
+            width = 0.0;
 
             if (!TryGetGlyphIndex(characterIdentifier, characterCodeToGlyphId, out var index))
             {
@@ -111,7 +110,7 @@
             return TableRegister.HeaderTable.UnitsPerEm;
         }
 
-        private bool TryGetBoundingAdvancedWidthByIndex(int index, out decimal width)
+        private bool TryGetBoundingAdvancedWidthByIndex(int index, out double width)
         {
             width = 0;
